moved whole calq stuff into calq dir
