+++ src/genie/CMakeLists.txt
@@ -17,4 +17,3 @@ add_subdirectory(util)
 add_subdirectory(quality/qvwriteout)
 add_subdirectory(name/tokenizer)
 add_subdirectory(quality/calq)
-add_subdirectory(quality/qvcalq)
