+++ src/genie/quality/calq/encoder.cc
@@ -4,12 +4,12 @@
  * https://github.com/mitogen/genie for more details.
  */
 
-#include "genie/quality/qvcalq/encoder.h"
+#include "genie/quality/calq/encoder.h"
 #include <string>
 #include <utility>
 #include "genie/core/record/alignment_split/same-rec.h"
 #include "genie/quality/calq/calq_coder.h"
-#include "genie/quality/qvcalq/permutation.h"
+#include "genie/quality/calq/permutation.h"
 #include "genie/util/watch.h"
 #include "iostream"
 // ---------------------------------------------------------------------------------------------------------------------
