+++ src/genie/quality/calq/permutation.h
